phase trans models now all differentiated - example now uses heat flux BC at the "1D heat conduction" components
this is important when using heat flox conditions at the ports (not temperatures). For heat flux conditions Dymola will try to differentiate the models, and then throws an error for the models defined at the BC/ for the ports (!). Seems that only the models at the ports are the poblem.